requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0
python-telegram-bot==20.7
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
aiohttp==3.9.1
python-dotenv==1.0.0
playwright-stealth==2.0.2